requirements.txt
pyrogram
tgcrypto
flask
gunicorn
Flask-SQLAlchemy
python-dotenv
psycopg2-binary